citation average updated to int instead of float. tc_list removed from final json

+++ PythonCode/Utilities/My_Data_Classes.py
@@ -14,8 +14,9 @@ class CategoryInfo:
     titles: Set[str] = field(default_factory=set)
     tc_count: int = 0
     tc_list: List[int] = field(default_factory=list)
-    citation_average: float = 0.0
+    citation_average: int = 0
     
+    # this holds the file names associated with articles
     # article_set: Set[str] = field(default_factory=set)
 
     def to_dict(self) -> dict:
